# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Enicarthage Staff Manager API",
    "version": "1.0.0",
    "description": "API pour la gestion des demandes d'impression à l'ENI Carthage"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "Endpoints pour l'authentification"
    },
    {
      "name": "Tickets",
      "description": "Endpoints pour la gestion des tickets d'impression"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Inscription d'un nouvel utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé avec succès"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Connexion utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Récupérer tous les tickets",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Liste des tickets"
          }
        }
      },
      "post": {
        "tags": ["Tickets"],
        "summary": "Créer un nouveau ticket",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "ticket": {
                    "$ref": "#/components/schemas/TicketCreateRequest"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket créé avec succès"
          }
        }
      }
    },
    "/api/tickets/{id}": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Récupérer un ticket par ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket trouvé"
          }
        }
      },
      "put": {
        "tags": ["Tickets"],
        "summary": "Mettre à jour un ticket",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket mis à jour"
          }
        }
      },
      "delete": {
        "tags": ["Tickets"],
        "summary": "Supprimer un ticket (Admin seulement)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket supprimé"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": {"type": "string"},
          "email": {"type": "string", "format": "email"},
          "password": {"type": "string", "minLength": 6},
          "department": {"type": "string"}
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {"type": "string", "format": "email"},
          "password": {"type": "string"}
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {"type": "string"},
          "type": {"type": "string", "default": "Bearer"},
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "email": {"type": "string"},
          "role": {"type": "string"}
        }
      },
      "TicketCreateRequest": {
        "type": "object",
        "required": ["title", "description", "actProf", "totalNumber"],
        "properties": {
          "title": {"type": "string", "minLength": 3, "maxLength": 200},
          "description": {"type": "string"},
          "actProf": {"type": "string"},
          "totalNumber": {"type": "integer", "minimum": 1, "maximum": 500}
        }
      },
      "TicketUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {"type": "string"},
          "description": {"type": "string"},
          "actProf": {"type": "string"},
          "totalNumber": {"type": "integer"},
          "status": {
            "type": "string",
            "enum": ["EN_ATTENTE", "EN_COURS", "TERMINE", "ANNULE"]
          }
        }
      }
    }
  }
}